refactor: move fastapi app to diff file
